refactor: redesign abstract repository function structure for product domain
